use memorypack instead of binaryformatter
